split current action and fmt macros
